+++ api/v1/class/exports/modules/CSVExportModule/CSVExportModule.em.arbeit.inc.php
@@ -19,7 +19,7 @@ class CSVExportModule implements ExportModuleInterface
         $month = $args["month"];
         $user = $args["user"];
 
-        
+        $this->saveAsCsv($args);
         if (!is_string($year)) {
             $year = date("Y");
         }
@@ -57,6 +57,56 @@ class CSVExportModule implements ExportModuleInterface
         fclose($output);
         exit;
     }
+
+    public function saveAsCsv($args) {
+        $arbeit = new Arbeitszeit();
+        $year = $args["year"];
+        $month = $args["month"];
+        $user = $args["user"];
+    
+        if (!is_string($year)) {
+            $year = date("Y");
+        }
+    
+        $sql = "SELECT id, username, schicht_tag, schicht_anfang, schicht_ende, ort, pause_start, pause_end 
+                FROM `arbeitszeiten` 
+                WHERE YEAR(schicht_tag) = ? AND MONTH(schicht_tag) = ? AND username = ? 
+                ORDER BY schicht_tag DESC";
+        $statement = $arbeit->db()->sendQuery($sql);
+        $statement->execute([$year, $month, $user]);
+        $data = $statement->fetchAll(\PDO::FETCH_ASSOC);
+    
+        // Falls keine Daten vorhanden sind
+        if (empty($data)) {
+            return false;
+        }
+    
+        // Erstelle den Dateipfad basierend auf den Argumenten
+        $directory = $_SERVER["DOCUMENT_ROOT"] . "/data/exports/" . $this->getName() . "/$user";
+        $filename = "$directory/worktimes_{$year}-{$month}.csv";
+    
+        // Verzeichnis erstellen, falls es nicht existiert
+        if (!is_dir($directory)) {
+            mkdir($directory, 0777, true);
+        }
+    
+        // Datei öffnen
+        $output = fopen($filename, 'w');
+    
+        // Setze die Spaltennamen in die CSV-Datei
+        $columns = ["ID", "Username", "Shift Date", "Shift Start", "Shift End", "Location/Notes", "Pause Start", "Pause End"];
+        fputcsv($output, $columns, ';');
+    
+        // Datenzeilen in CSV schreiben
+        foreach ($data as $row) {
+            fputcsv($output, $row, ';');
+        }
+    
+        fclose($output);
+    
+        return $filename;
+    }
+    
 
     public function getName() {
         return "CSVExportModule";