+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Added CSV export module class `Arbeitszeit\ExportModule\CSVExportModule` which can be used by clicking on `(CSV)` within All Worktimes
 
+<!-- Fixed title not set for Plugin Hub -->
+
 ## v7.4
 
 * Exports are now handled differently. You can import your own modules to support more file formats. Read more about this in the `README.md` file.