+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Added CSV export module class `Arbeitszeit\ExportModule\CSVExportModule` which can be used by clicking on `(CSV)` within All Worktimes
 * Renamed all GUI elements from `Calendar` to `Notifications`
+* Fixed being unable to edit notifications entries
 
 <!-- Fixed title not set for Plugin Hub -->