+++ CHANGELOG.md
@@ -3,9 +3,12 @@
 ## v7.5
 
 * Added CSV export module class `Arbeitszeit\ExportModule\CSVExportModule` which can be used by clicking on `(CSV)` within All Worktimes
+* Renamed all GUI elements from `Calendar` to `Notifications`
+* Fixed being unable to edit notifications entries
 * PDF and CSV exports are now directly saved onto the server. This is done automatically. Exports are saved within `data/exports/{ExportModuleName}/{username}/`
 
 <!-- Fixed title not set for Plugin Hub -->
+<!-- Fixed missing i18n files for `suite/admin/notifications/edit.php` -->
 
 ## v7.4